Refactor search results related components in their own folders
Don't show in changelog
