+++ newIDE/app/src/AssetStore/ExtensionStore/ExtensionStoreContext.js
@@ -1,13 +1,13 @@
 // @flow
 import * as React from 'react';
-import { type FiltersState, useFilters } from '../FiltersChooser';
+import { type FiltersState, useFilters } from '../../UI/Search/FiltersChooser';
 import { type Filters } from '../../Utils/GDevelopServices/Asset';
 import {
   getExtensionsRegistry,
   type ExtensionsRegistry,
   type ExtensionShortHeader,
 } from '../../Utils/GDevelopServices/Extension';
-import { useSearchItem } from '../UseSearchItem';
+import { useSearchItem } from '../../UI/Search/UseSearchItem';
 
 const defaultSearchText = '';
