+++ newIDE/app/src/AssetStore/ExtensionStore/index.js
@@ -5,9 +5,9 @@ import { Column, Line } from '../../UI/Grid';
 import Background from '../../UI/Background';
 import ScrollView from '../../UI/ScrollView';
 import { type ExtensionShortHeader } from '../../Utils/GDevelopServices/Extension';
-import { FiltersChooser } from '../FiltersChooser';
+import { FiltersChooser } from '../../UI/Search/FiltersChooser';
 import { ExtensionStoreContext } from './ExtensionStoreContext';
-import { ListSearchResults } from './ListSearchResults';
+import { ListSearchResults } from '../../UI/Search/ListSearchResults';
 import { ExtensionListItem } from './ExtensionListItem';
 import { ResponsiveWindowMeasurer } from '../../UI/Reponsive/ResponsiveWindowMeasurer';
 import ExtensionInstallDialog from './ExtensionInstallDialog';
