+++ newIDE/app/src/Utils/GDevelopServices/Types.flow.js
@@ -1,0 +1,13 @@
+// @flow
+
+export type TagsTreeNode = {|
+  name: string,
+  children: Array<TagsTreeNode>,
+  allChildrenTags: Array<string>,
+|};
+
+export type Filters = {|
+  allTags: Array<string>,
+  defaultTags: Array<string>,
+  tagsTree: Array<TagsTreeNode>,
+|};
